refactor: move threshold card as components
